+++ address-space-controller/src/main/java/io/enmasse/controller/TemplateInfraResourceFactory.java
@@ -100,11 +100,6 @@ public class TemplateInfraResourceFactory implements InfraResourceFactory {
             parameters.put(TemplateParameter.AUTHENTICATION_SERVICE_SASL_INIT_HOST, addressSpace.getAnnotation(AnnotationKeys.REALM_NAME));
         }
 
-        if (authService.getMetadata().getAnnotations() != null &&
-                authService.getMetadata().getAnnotations().get(AnnotationKeys.OAUTH_URL) != null) {
-            parameters.put(TemplateParameter.AUTHENTICATION_SERVICE_OAUTH_URL, authService.getMetadata().getAnnotations().get(AnnotationKeys.OAUTH_URL));
-        }
-
         Map<String, CertSpec> serviceCertMapping = new HashMap<>();
         for (EndpointSpec endpoint : addressSpace.getSpec().getEndpoints()) {
             if (endpoint.getCert() != null) {