+++ address-space-controller/src/main/java/io/enmasse/controller/keycloak/RealmController.java
@@ -6,71 +6,26 @@
 package io.enmasse.controller.keycloak;
 
 import io.enmasse.address.model.AddressSpace;
-import io.enmasse.address.model.EndpointSpec;
-import io.enmasse.address.model.EndpointStatus;
-import io.enmasse.address.model.KubeUtil;
 import io.enmasse.admin.model.v1.AuthenticationService;
 import io.enmasse.admin.model.v1.AuthenticationServiceType;
 import io.enmasse.config.AnnotationKeys;
 import io.enmasse.controller.Controller;
 import io.enmasse.k8s.api.AuthenticationServiceRegistry;
-import io.enmasse.user.api.UserApi;
-import io.enmasse.user.model.v1.*;
-import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static io.enmasse.user.model.v1.Operation.*;
-
 public class RealmController implements Controller {
     private static final Logger log = LoggerFactory.getLogger(RealmController.class);
     private static final String MASTER_REALM = "master";
     private final KeycloakApi keycloak;
-    private final UserLookupApi userLookupApi;
-    private final UserApi userApi;
     private final AuthenticationServiceRegistry authenticationServiceRegistry;
-    private final Map<String, KeycloakRealmParams> lastParams = new HashMap<>();
 
-    public RealmController(KeycloakApi keycloak, UserLookupApi userLookupApi, UserApi userApi, AuthenticationServiceRegistry authenticationServiceRegistry) {
+    public RealmController(KeycloakApi keycloak, AuthenticationServiceRegistry authenticationServiceRegistry) {
         this.keycloak = keycloak;
-        this.userLookupApi = userLookupApi;
-        this.userApi = userApi;
         this.authenticationServiceRegistry = authenticationServiceRegistry;
-    }
-
-    private EndpointSpec getConsoleEndpoint(AddressSpace addressSpace) {
-        for (EndpointSpec endpoint : addressSpace.getSpec().getEndpoints()) {
-            if (endpoint.getService().startsWith("console")) {
-                return endpoint;
-            }
-        }
-        return null;
-    }
-
-    private EndpointStatus getConsoleEndpointStatus(AddressSpace addressSpace) {
-        EndpointSpec spec = getConsoleEndpoint(addressSpace);
-        if (spec == null) {
-            return null;
-        }
-
-        for (EndpointStatus endpoint : addressSpace.getStatus().getEndpointStatuses()) {
-            if (endpoint.getName().equals(spec.getName())) {
-                return endpoint;
-            }
-        }
-        return null;
-    }
-
-    private String getConsoleUri(EndpointStatus endpoint) {
-        String uri = null;
-        if (endpoint.getExternalHost() != null) {
-            uri = "https://" + endpoint.getExternalHost() + "/*";
-            log.info("Using {} as redirect URI for enmasse-console", uri);
-        }
-        return uri;
     }
 
     private static class AuthServiceEntry {
@@ -114,11 +69,6 @@ public class RealmController implements Controller {
                 continue;
             }
 
-            if (addressSpace.getSpec().getEndpoints() == null) {
-                log.debug("No endpoints defined for address space, not performing any operations");
-                continue;
-            }
-
             AuthServiceEntry entry = authserviceMap.computeIfAbsent(authenticationService.getMetadata().getName(), k -> new AuthServiceEntry(authenticationService));
             entry.addAddressSpace(addressSpace);
         }
@@ -143,12 +93,6 @@ public class RealmController implements Controller {
         for (AuthServiceEntry entry : authserviceMap.values()) {
             AuthenticationService authenticationService = entry.getAuthenticationService();
             List<AddressSpace> addressSpaces = entry.getAddressSpaces();
-            KeycloakRealmParams keycloakRealmParams = KeycloakRealmParams.fromAuthenticationService(authenticationService);
-            if (!Objects.equals(lastParams.computeIfAbsent(authenticationService.getMetadata().getName(), k -> KeycloakRealmParams.NULL_PARAMS), keycloakRealmParams)) {
-                log.info("Identity provider params: {}", keycloakRealmParams);
-                updateExistingRealms(authenticationService, keycloakRealmParams);
-                lastParams.put(authenticationService.getMetadata().getName(), keycloakRealmParams);
-            }
 
             Set<String> actualRealms = keycloak.getRealmNames(authenticationService);
 
@@ -158,54 +102,9 @@ public class RealmController implements Controller {
                     continue;
                 }
                 log.info("Creating realm {} in authentication service {}", realmName, authenticationService.getMetadata().getName());
-                String userName = addressSpace.getAnnotation(AnnotationKeys.CREATED_BY);
-                String userId = addressSpace.getAnnotation(AnnotationKeys.CREATED_BY_UID);
-                if (userId == null || userId.isEmpty()) {
-                    userId = userLookupApi.findUserId(userName);
-                }
-
-                EndpointStatus endpointStatus = getConsoleEndpointStatus(addressSpace);
-                if (endpointStatus == null) {
-                    log.info("Address space {} has no endpoints defined", addressSpace.getMetadata().getName());
-                } else if (endpointStatus.getExternalHost() == null && endpointStatus.getExternalPorts().isEmpty()) {
-                    log.info("Address space {} console endpoint host not known, waiting", addressSpace.getMetadata().getName());
-                } else {
-                    String consoleUri = getConsoleUri(endpointStatus);
-                    keycloak.createRealm(authenticationService, addressSpace.getMetadata().getNamespace(), realmName, consoleUri, keycloakRealmParams);
-                    userApi.createUser(authenticationService, realmName, new UserBuilder()
-                            .withMetadata(new ObjectMetaBuilder()
-                                    .withName(addressSpace.getMetadata().getName() + "." + KubeUtil.sanitizeUserName(userName))
-                                    .withNamespace(addressSpace.getMetadata().getNamespace())
-                                    .build())
-                            .withSpec(new UserSpecBuilder()
-                                    .withUsername(KubeUtil.sanitizeUserName(userName))
-                                    .withAuthentication(new UserAuthenticationBuilder()
-                                            .withType(UserAuthenticationType.federated)
-                                            .withProvider("openshift")
-                                            .withFederatedUserid(userId)
-                                            .withFederatedUsername(userName)
-                                            .build())
-                                    .withAuthorization(Arrays.asList(new UserAuthorizationBuilder()
-                                                    .withAddresses(Collections.singletonList("*"))
-                                                    .withOperations(Arrays.asList(send, recv, view))
-                                                    .build(),
-                                            new UserAuthorizationBuilder()
-                                                    .withOperations(Collections.singletonList(Operation.manage))
-                                                    .build()))
-                                    .build())
-                            .build());
-                }
+                keycloak.createRealm(authenticationService, addressSpace.getMetadata().getNamespace(), realmName);
             }
         }
-    }
-
-    private void updateExistingRealms(AuthenticationService authenticationService, KeycloakRealmParams updatedParams) {
-        keycloak.getRealmNames(authenticationService).stream()
-                .filter(name -> !name.equals(MASTER_REALM))
-                .forEach(name -> {
-                    log.info("Updating realm {}", name);
-                    keycloak.updateRealm(authenticationService, name, updatedParams);
-                });
     }
 
     @Override