+++ address-space-controller/src/main/java/io/enmasse/controller/keycloak/UserLookupApi.java
@@ -1,9 +1,0 @@
-/*
- * Copyright 2017-2018, EnMasse authors.
- * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
- */
-package io.enmasse.controller.keycloak;
-
-public interface UserLookupApi {
-    String findUserId(String userName);
-}